\name{anovaTableI.HH}
\alias{anovaTableI.HH}
\alias{anovaTableII.HH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rcmdr interfacce to anova function}
\description{
  Rcmdr interface to \code{anova} function, specifically to get the
  sequential sums of squares.
}
\usage{
anovaTableI.HH()
anovaTableII.HH()  ## exact copy of John Fox's anovaTable from Rcmdr/R/model-menu.R
}
%- maybe also 'usage' for other objects documented here.
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{ \code{\link[stats]{anova}}}
\keyword{ regression }
