\name{BoxCox}
\alias{BoxCox}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Rcmdr BoxCox demo renamed to active function. }
\description{
  Rcmdr menu for Box-Cox Transformations
}
\usage{
BoxCox()
}
%- maybe also 'usage' for other objects documented here.
\author{John Fox \email{jfox@mcmaster.ca}}
\keyword{manip}
