\name{confidenceIntervalsPlot}
\alias{confidenceIntervalsPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rcmdr interface to plot confidence and prediction intervals in
  simple linear regression}
\description{
  Rcmdr menu interface to the function \code{ci.plot}.
  Variable boxes are provided
  for one predictor variable, one response variable.
  The simple linear regression is calculated and made the
  active model.
}
\usage{
confidenceIntervalsPlot()
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{\code{\link[HH]{ci.plot}}}
\keyword{hplot}
