\name{DotplottbRcmdr}
\alias{DotplottbRcmdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rcmdr menu interface to dotplot(panel=panel.dotplot.tb).}
\description{
Rcmdr menu interface to dotplot(panel=panel.dotplot.tb).
}
\usage{
DotplottbRcmdr()
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{\code{\link[HH]{panel.dotplot.tb}}}
\keyword{hplot}
