\name{Interaction2wtRcmdr}
\alias{Interaction2wtRcmdr}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rcmdr menu interface to interaction2wt}
\description{
  Plot all main effects and twoway interactions in a multifactor design.
The main diagonal
displays boxplots for the main effects of each factor.  The
off-diagonals show the interaction plots for each pair of factors.
The \code{i,j} panel shows the same factors as the \code{j,i} but with
the trace- and x-factor roles interchanged.
  
}
\usage{
Interaction2wtRcmdr()
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{\code{\link[HH]{interaction2wt}}}
\keyword{hplot}
