\name{MMCmenu}
\alias{MMCmenu}
\alias{MMC2menu}
\alias{AOVModelsP}
\title{Menu interface to MMC plots.}
\description{Menu interface to MMC (Mean--mean Multiple Comparison) plots.}

\usage{
MMCmenu()
MMC2menu()
AOVModelsP(n=1)
}
\arguments{
  \item{n}{Minimum number of \code{"aov"} models.}
  }
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{hplot}
