\name{normalHypothesesPlot}
\alias{normalHypothesesPlot}
\alias{tHypothesesPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Rcmdr normalHypothesesPlot and tHypothesesPlot menu.}
\description{
  Rcmdr normalHypothesesPlot and tHypothesesPlot menu
}
\usage{
normalHypothesesPlot()

tHypothesesPlot()
}
%- maybe also 'usage' for other objects documented here.
\author{Richard M. Heiberger <rmh@temple.edu>.}
\seealso{\code{\link[HH]{norm.curve}}}
\keyword{hplot}
