\name{Projector}
\alias{Projector}
\alias{H600}
\title{Set Rcmdr options for good visibility on classroom projector and
  on netbook screen with 600 pixel height.}
\description{Set Rcmdr options for good visibility on classroom projector and
  on netbook screen with 600 pixel height.}

\usage{
Projector()
H600()
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{environment}
