\name{Regr1Plot}
\alias{Regr1Plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rcmdr Menu function to display the squared residuals.}
\description{
  Rcmdr Menu function to display the squared residuals of a linear fit
  of one y variable on one x variable.  The default model is simple
  linear regression \code{y ~ x}.  Any other model of one y on one x may
  be used.  See the last example in \code{\link[HH]{regr1.plot}} for an
  example of a quadratic function of x.
}
\usage{
Regr1Plot()
}
%- maybe also 'usage' for other objects documented here.
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{\code{\link[HH]{regr1.plot}}}
\keyword{hplot}
