\name{scatterPlot.HH}
\alias{scatterPlot.HH}
\title{Scatterplot menu with different defaults than Rcmdr.}
\description{Alternate menu into the scatterplot in the car package.
  This menu by default uses solid dots, larger fonts, and turns off marginal boxplots
  and smoother lines.  Otherwise it is identical to the Rcmdr Scatterplot menu
  item.}

\usage{
scatterPlot.HH()
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{hplot}
