\name{scatterPlotMatrix.HH}
\alias{scatterPlotMatrix.HH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Scatterplot Matrices}
\description{
  This is variation of the \code{Rcmdr} interface to the \code{car}
  package \code{scatterplot.matrix}.
  The revision uses \code{row1attop=FALSE} to force the main diagonal of
  the scatterplot matrix to go uphill from southwest to northeast.
}
\usage{
scatterPlotMatrix.HH()
}
\author{John Fox \email{jfox@mcmaster.ca}.
  \code{row1attop=FALSE} added by Richard M. Heiberger <rmh@temple.edu>.
}
\seealso{\code{\link[car]{scatterplot.matrix}}}
\keyword{hplot}
