\name{normal.and.t.hypotheses.plot}
\alias{normal.and.t.hypotheses.plot}
\alias{normalHypothesesPlot}
\alias{tHypothesesPlot}
\alias{FHypothesesPlot}
\alias{ChisqHypothesesPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Rcmdr normalHypothesesPlot and tHypothesesPlot menu.}
\description{
  Rcmdr menus to draw graphs of hypotheses, critical values, and p-values.
}
\usage{
normal.and.t.hypotheses.plot()

%normalHypothesesPlot() ## deprecated in favor of normal.and.t.hypotheses.plot()

%tHypothesesPlot()      ## deprecated in favor of normal.and.t.hypotheses.plot()

FHypothesesPlot()

ChisqHypothesesPlot()
}
%- maybe also 'usage' for other objects documented here.
\author{Richard M. Heiberger <rmh@temple.edu>.}
\seealso{\code{\link[HH]{norm.curve}}, \code{\link[HH]{F.curve}}, \code{\link[HH:F.curve]{chisq.curve}}}
\keyword{hplot}
