\name{normalHypothesesPlot}
\alias{normalHypothesesPlot}
\alias{tHypothesesPlot}
\alias{FHypothesesPlot}
\alias{ChisqHypothesesPlot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Rcmdr normalHypothesesPlot and tHypothesesPlot menu.}
\description{
  Rcmdr menus to draw graphs of hypotheses, critical values, and p-values.
}
\usage{
normalHypothesesPlot()

tHypothesesPlot()

FHypothesesPlot()

ChisqHypothesesPlot()
}
%- maybe also 'usage' for other objects documented here.
\author{Richard M. Heiberger <rmh@temple.edu>.}
\seealso{\code{\link[HH]{norm.curve}}, \code{\link[HH]{F.curve}}, \code{\link[HH]{chisq.curve}}}
\keyword{hplot}
