\name{scatterPlot.HH}
\alias{scatterPlot.HH}
\title{scatterplot menu with different defaults}
\description{Alternate menu into the scatterplot in the car package.
  This menu by default uses solid dots and turns off marginal boxplots
  and smoother lines.  Otherwise it is identical to the Rcmdr menu
  item.}

\usage{
scatterPlot.HH()
}
\author{ Richard M. Heiberger <rmh@temple.edu> }
\keyword{hplot}
