\name{Xyplot.HH}
\alias{Xyplot.HH}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Rcmdr Menu function to specify an xyplot.}
\description{
This is an enhancement of the Rcmdr {\tt Xyplot} function (which I
wrote) to include layout parameters, to force solid dots, and to
distinguish
between conditioning variables in the formula and group varibles.
}
\usage{
Xyplot.HH()
}
%- maybe also 'usage' for other objects documented here.
\author{ Richard M. Heiberger <rmh@temple.edu> }
\seealso{\code{\link[lattice]{xyplot}}}
\keyword{hplot}
